refactoring unit test internals
